Moved mx-pixels tests to TestImageAppParameter class
